backend/requirements.txt
Flask==2.3.4
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.1
google-generative-ai==0.11.0
pymongo==4.4.0
